reminders will no longer duplicate
