--- lambda/custom/custom/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "hello-world",
-  "version": "1.1.0",
-  "description": "alexa utility for quickly building skills",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Amazon Alexa",
-  "license": "ISC",
-  "dependencies": {
-    "ask-sdk-core": "^2.6.0",
-    "ask-sdk-model": "^1.18.0",
-    "aws-sdk": "^2.326.0",
-    "axios": "^0.19.0"
-  }
-}
