+++ lambda/index.js
@@ -1,8 +1,9 @@
 const Alexa = require("ask-sdk-core");
-const API = require('./util/apiUtils')
-const baseUrl = "https://medirep-api.herokuapp.com/api";
 const axios = require("axios"); //TODO: extract axios requests to util functions
+const baseUrl = "https://medirep-api.herokuapp.com/api";
 const user_id = 1;
+const PERMISSIONS = ['alexa::alerts:reminders:skill:readwrite'];
+const utils = require('./utils/Utils')
 
 const LaunchRequestHandler = {
   canHandle(handlerInput) {
@@ -11,8 +12,18 @@ const LaunchRequestHandler = {
     );
   },
   handle(handlerInput) {
+    const { responseBuilder } = handlerInput
+    const requestEnvelope = handlerInput.requestEnvelope;
+    const permissions = requestEnvelope.context.System.user.permissions
+      if (!permissions) {
+        // if no permissions, nag the user to grant them
+        return responseBuilder
+          .speak('I need permission to create reminders. Take a look at your alexa app to grant them.')
+          .withAskForPermissionsConsentCard(["alexa::alerts:reminders:skill:readwrite"])
+          .getResponse()
+      }
     const speakOutput = "what would you like me to do";
-    return handlerInput.responseBuilder
+    return responseBuilder
       .speak(speakOutput)
       .reprompt(speakOutput)
       .getResponse();
@@ -64,7 +75,6 @@ const PairDeviceIntentHandler = {
     const pairDeviceCode =
       handlerInput.requestEnvelope.request.intent.slots.pairDeviceCode.value;
     const amazon_id = handlerInput.requestEnvelope.session.user.userId;
-    console.log(amazon_id);
     return handlerInput.responseBuilder.speak(pairDeviceCode).getResponse();
   }
 };
@@ -74,14 +84,20 @@ const newReminderIntentHandler = {
       return Alexa.getRequestType(handlerInput.requestEnvelope) === 'IntentRequest'
           && Alexa.getIntentName(handlerInput.requestEnvelope) === 'newReminderIntent';
   },
-
   async handle(handlerInput) {
     // TODO: This needs to read the medication list and make sure that reminders for those items exist in the amazon reminders API
       const requestEnvelope = handlerInput.requestEnvelope;
       const responseBuilder = handlerInput.responseBuilder;
-      const consentToken = requestEnvelope.context.System.apiAccessToken;
+      const permissions = requestEnvelope.context.System.user.permissions
+      if (!permissions) {
+        // if no permissions, nag the user to grant them
+        return responseBuilder
+          .speak('I need permission to create reminders. Take a look at your alexa app to grant them.')
+          .withAskForPermissionsConsentCard(PERMISSIONS)
+          .getResponse()
+      }
       switch (requestEnvelope.request.intent.confirmationStatus) {
-              case 'CONFIRMED':getAmazon
+              case 'CONFIRMED':
                 console.log('Alexa confirmed intent, so clear to create reminder');
                 break;
               case 'DENIED':
@@ -97,57 +113,32 @@ const newReminderIntentHandler = {
               .addDelegateDirective()
               .getResponse();
         }
-
-      if (!consentToken) {
-          return responseBuilder
-          .speak("you don't currently have reminders enabled for this skill")
-            .withAskForPermissionsConsentCard(PERMISSIONS)
-            .getResponse();
-        }
         try {
-            
             const client = handlerInput.serviceClientFactory.getReminderManagementServiceClient();
-            
-
-
-            const reminderRequest = {
-                trigger: {
-                    type: 'SCHEDULED_RELATIVE',
-                    offsetInSeconds: '5',
-                  },
-                  alertInfo: {
-                      spokenInfo: {
-                          content: [{
-                              locale: 'en-GB',
-                              text: `Take your pills!`,
-                          }],
-                      },  
-                  },
-                  pushNotification: {
-                      status: 'ENABLED',
-                  },
-              };
-          const reminderResponse = await client.createReminder(reminderRequest).then(console.log);
+            axios.get('https://medirep-api.herokuapp.com/api/meds/1')
+            .then( async ({data: {meds}}) => {
+              const medsReminders = utils.reminderBuilder(meds)
+              const { alerts: presentReminders } = await client.getReminders();
+              const filteredReminders = utils.filterAgainstExistingReminders(presentReminders, medsReminders)
+              console.log(filteredReminders)
+              filteredReminders.forEach(async reminder => {
+                await client.createReminder(reminder).then(console.log);
+              })
+            })
+            .catch(console.log)
       } catch (error) {
               if (error.name !== 'ServiceError') {
-              console.log(`error: ${error.stack}`);
-              const response = responseBuilder.speak('Something Went Terribly Wrong').getResponse();
-              return response;
+                console.log(`error: ${error.stack}`);
+                const response = responseBuilder.speak(`Something went wrong with the reminders`).getResponse();
+                return response;
               }
-          throw error;
-      }
-      // TODO: Get the amazon id from the request header here.
-      const currentReminders = API.getRemindersFromAmazon(user_id)
-
-      return responseBuilder.speak(`you now have ${currentReminders.length} reminders`)
-          .getResponse()
+              throw error;
+            }
+    return responseBuilder
+      .speak(`Reminders are up to date`)
+      .getResponse()
   }
 };
-
-
-
-
-
 
 const HelpIntentHandler = {
   canHandle(handlerInput) {
@@ -158,13 +149,13 @@ const HelpIntentHandler = {
   },
   handle(handlerInput) {
     const speakOutput = "You can say hello to me! How can I help?";
-
     return handlerInput.responseBuilder
       .speak(speakOutput)
       .reprompt(speakOutput)
       .getResponse();
   }
 };
+
 const CancelAndStopIntentHandler = {
   canHandle(handlerInput) {
     return (
@@ -180,6 +171,7 @@ const CancelAndStopIntentHandler = {
     return handlerInput.responseBuilder.speak(speakOutput).getResponse();
   }
 };
+
 const SessionEndedRequestHandler = {
   canHandle(handlerInput) {
     return (
@@ -191,6 +183,7 @@ const SessionEndedRequestHandler = {
     return handlerInput.responseBuilder.getResponse();
   }
 };
+
 const IntentReflectorHandler = {
   canHandle(handlerInput) {
     return (
@@ -200,7 +193,6 @@ const IntentReflectorHandler = {
   handle(handlerInput) {
     const intentName = Alexa.getIntentName(handlerInput.requestEnvelope);
     const speakOutput = `You just triggered ${intentName}`;
-
     return handlerInput.responseBuilder.speak(speakOutput).getResponse();
   }
 };
@@ -219,16 +211,32 @@ const ErrorHandler = {
   }
 };
 
+const RequestLog = {
+  async process(handlerInput) {
+    console.log(`REQUEST ENVELOPE = ${JSON.stringify(handlerInput.requestEnvelope)}`);
+  },
+};
+  
+const ResponseLog = {
+  process(handlerInput) {
+    console.log(`RESPONSE = ${JSON.stringify(handlerInput.responseBuilder.getResponse())}`);
+   },
+};
+
 exports.handler = Alexa.SkillBuilders.custom()
   .addRequestHandlers(
     LaunchRequestHandler,
     QuizIntentHandler,
     PairDeviceIntentHandler,
+    newReminderIntentHandler,
     HelpIntentHandler,
     CancelAndStopIntentHandler,
     SessionEndedRequestHandler,
-    IntentReflectorHandler,
-    newReminderIntentHandler
+    IntentReflectorHandler
   )
+  .withApiClient(new Alexa.DefaultApiClient())
   .addErrorHandlers(ErrorHandler)
+  .addRequestInterceptors(RequestLog)
+  .addResponseInterceptors(ResponseLog)
+  .withCustomUserAgent(`Hayley_smells/v1`)
   .lambda();
