+++ lambda/util.js
